# If the poly has 1 evaluation, we should fix the window size to 1