@mmagician 2023-11-02
If only all my optimizations were as significant as this one 😆